Other version2.0

+++ EntityLayer/Abstract/IKullanici.cs
@@ -15,12 +15,12 @@ namespace EntityLayer.Abstract
         string Soyadı { get; set; }
         string KullanıcıMail { get; set; }
         string KullanıcıŞifre { get; set; }
-        int Boy { get; set; }
+        double Boy { get; set; }
         Cinsiyet Cinsiyet { get; set; }
         //Bir kullanıcı birden fazla besin tüketecek
         //List<TuketilenBesin> TüketilenBesinler { get; set; }
         int Yas { get; set; }
-        int MevcutAğırlık { get; set; }
+        double MevcutAğırlık { get; set; }
         string AktiviteDüzeyi { get; set; }
         string DiyetHedefi { get; set; }
         int HedefAgırlıgı { get; set; }
